chore: move `validateInputUsingValidator` to `input_schema` package
