+++ packages/input_schema/src/utilities.ts
@@ -1,0 +1,248 @@
+import { PROXY_URL_REGEX, URL_REGEX } from '@apify/consts';
+import { countries } from 'countries-list';
+import { ValidateFunction } from 'ajv';
+import { m } from './intl';
+import { parseAjvError } from './input_schema';
+
+/**
+ * Validate's input field configured with proxy editor
+ * @param fieldKey Proxy field value
+ * @param value Proxy field value
+ * @param [isRequired] Whether the field is required or not
+ * @param [options] Information about proxy groups availability
+ * @param [options.hasAutoProxyGroups] Informs validation whether user has atleast one proxy group available in auto mode
+ * @param [options.availableProxyGroups] List of available proxy groups
+ * @param [options.disabledProxyGroups] Object with groupId as key and error message as value (mostly for residential/SERP)
+ */
+function validateProxyField(
+    fieldKey: Record<string, unknown>,
+    value: Record<string, any>,
+    isRequired = false,
+    options: { hasAutoProxyGroups?: boolean; availableProxyGroups?: string[]; disabledProxyGroups?: Record<string, unknown> } | null = null,
+) {
+    const fieldErrors: any[] = [];
+    if (isRequired) {
+        // Nullable error is already handled by AJV
+        if (value === null) return fieldErrors;
+        if (!value) {
+            const message = m('inputSchema.validation.required', { rootName: 'input', fieldKey });
+            fieldErrors.push(message);
+            return fieldErrors;
+        }
+
+        const { useApifyProxy, proxyUrls } = value;
+        if (!useApifyProxy && (!Array.isArray(proxyUrls) || proxyUrls.length === 0)) {
+            fieldErrors.push(m('inputSchema.validation.proxyRequired', { rootName: 'input', fieldKey }));
+            return fieldErrors;
+        }
+    }
+
+    // Input is not required, so missing value is valid
+    if (!value) return fieldErrors;
+
+    const { useApifyProxy, proxyUrls, apifyProxyGroups, apifyProxyCountry } = value;
+
+    if (!useApifyProxy && Array.isArray(proxyUrls)) {
+        let invalidUrl = false;
+        proxyUrls.forEach((url) => {
+            if (!PROXY_URL_REGEX.test(url.trim())) invalidUrl = url.trim();
+        });
+        if (invalidUrl) {
+            fieldErrors.push(m('inputSchema.validation.customProxyInvalid', { invalidUrl }));
+        }
+    }
+
+    // Apify proxy country can be set only when using Apify proxy
+    if (!useApifyProxy && apifyProxyCountry) {
+        fieldErrors.push(m('inputSchema.validation.apifyProxyCountryWithoutApifyProxyForbidden'));
+    }
+
+    // If Apify proxy is not used skip additional checks
+    if (!useApifyProxy) return fieldErrors;
+
+    // If Apify proxy is used, check if there is a selected country and if so, check that it's valid (empty or a valid country code)
+    if (apifyProxyCountry && !countries[apifyProxyCountry]) {
+        fieldErrors.push(m('inputSchema.validation.apifyProxyCountryInvalid', { invalidCountry: apifyProxyCountry }));
+    }
+
+    // If options are not provided skip additional checks
+    if (!options) return fieldErrors;
+
+    const selectedProxyGroups = (apifyProxyGroups || []);
+
+    // Auto mode, check that user has access to alteast one proxy group usable in this mode
+    if (!selectedProxyGroups.length && !options.hasAutoProxyGroups) {
+        fieldErrors.push(m('inputSchema.validation.noAvailableAutoProxy'));
+        return fieldErrors;
+    }
+
+    // Check if proxy groups selected by user are available to him
+    const availableProxyGroupsById = {};
+    (options.availableProxyGroups || []).forEach((group) => { availableProxyGroupsById[group] = true; });
+    const unavailableProxyGroups = selectedProxyGroups.filter((group: string) => !availableProxyGroupsById[group]);
+
+    if (unavailableProxyGroups.length) {
+        fieldErrors.push(m('inputSchema.validation.proxyGroupsNotAvailable', {
+            rootName: 'input',
+            fieldKey,
+            groups: unavailableProxyGroups.join(', '),
+        }));
+    }
+
+    // Check if any of the proxy groups are blocked and if yes then output the associated message
+    const blockedProxyGroupsById = options.disabledProxyGroups || {};
+    selectedProxyGroups
+        .filter((group: string) => blockedProxyGroupsById[group])
+        .forEach((blockedGroup: string) => {
+            fieldErrors.push(blockedProxyGroupsById[blockedGroup]);
+        });
+
+    return fieldErrors;
+}
+
+/**
+ * Uses AJV validator to validate input with input schema and then
+ * does custom validation for our own properties (nullable, patternKey, patternValue)
+ * @param validator Initialized AJV validator
+ * @param inputSchema Valid input schema in object
+ * @param input Input object to be validated
+ * @param options (Optional) Additional validation configuration for certain fields
+ */
+export function validateInputUsingValidator(
+    validator: ValidateFunction,
+    inputSchema: Record<string, any>,
+    input: Record<string, unknown>,
+    options: Record<string, any> = {},
+) {
+    const isValid = validator(input); // Check if input is valid based on schema values
+
+    const { properties } = inputSchema;
+    const required = inputSchema.required || [];
+
+    let errors: { fieldKey: string, message: string }[] = [];
+    // Process AJV validation errors
+    if (!isValid) {
+        errors = validator.errors!
+            .map((error) => parseAjvError(error, 'input', properties, input))
+            .filter((error) => !!error) as any[];
+    }
+
+    Object.keys(properties).forEach((property) => {
+        const value = input[property] as Record<string, any>;
+        const { type, editor, patternKey, patternValue } = properties[property];
+        const fieldErrors = [];
+        // Check that proxy is required, if yes, valides that it's correctly setup
+        if (type === 'object' && editor === 'proxy') {
+            const proxyValidationErrors = validateProxyField(property as any, value, required.includes(property), options.proxy);
+            proxyValidationErrors.forEach((error) => {
+                fieldErrors.push(error);
+            });
+        }
+        // Check that array items fit patternKey and patternValue
+        if (type === 'array' && value && Array.isArray(value)) {
+            if (editor === 'requestListSources') {
+                const invalidIndexes: any[] = [];
+                value.forEach((item, index) => {
+                    if (!item) invalidIndexes.push(index);
+                    else if (!item.url && !item.requestsFromUrl) invalidIndexes.push(index);
+                    else if (item.url && !URL_REGEX.test(item.url)) invalidIndexes.push(index);
+                    else if (item.requestsFromUrl && !URL_REGEX.test(item.requestsFromUrl)) invalidIndexes.push(index);
+                });
+                if (invalidIndexes.length) {
+                    fieldErrors.push(m('inputSchema.validation.requestListSourcesInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidIndexes: invalidIndexes.join(','),
+                    }));
+                }
+            }
+            // If patternKey is provided, then validate keys of objects in array
+            if (patternKey && editor === 'keyValue') {
+                const check = new RegExp(patternKey);
+                const invalidIndexes: any[] = [];
+                value.forEach((item, index) => {
+                    if (!check.test(item.key)) invalidIndexes.push(index);
+                });
+                if (invalidIndexes.length) {
+                    fieldErrors.push(m('inputSchema.validation.arrayKeysInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidIndexes: invalidIndexes.join(','),
+                        pattern: patternKey,
+                    }));
+                }
+            }
+            // If patternValue is provided and editor is keyValue, then validate values of objecs in array
+            if (patternValue && editor === 'keyValue') {
+                const check = new RegExp(patternValue);
+                const invalidIndexes: any[] = [];
+                value.forEach((item, index) => {
+                    if (!check.test(item.value)) invalidIndexes.push(index);
+                });
+                if (invalidIndexes.length) {
+                    fieldErrors.push(m('inputSchema.validation.arrayValuesInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidIndexes: invalidIndexes.join(','),
+                        pattern: patternValue,
+                    }));
+                }
+            // If patternValue is provided and editor is stringList, then validate each item in array
+            } else if (patternValue && editor === 'stringList') {
+                const check = new RegExp(patternValue);
+                const invalidIndexes: any[] = [];
+                value.forEach((item, index) => {
+                    if (!check.test(item)) invalidIndexes.push(index);
+                });
+                if (invalidIndexes.length) {
+                    fieldErrors.push(m('inputSchema.validation.arrayValuesInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidIndexes: invalidIndexes.join(','),
+                        pattern: patternValue,
+                    }));
+                }
+            }
+        }
+        // Check that object items fit patternKey and patternValue
+        if (type === 'object' && value) {
+            if (patternKey) {
+                const check = new RegExp(patternKey);
+                const invalidKeys: any[] = [];
+                Object.keys(value).forEach((key) => {
+                    if (!check.test(key)) invalidKeys.push(key);
+                });
+                if (invalidKeys.length) {
+                    fieldErrors.push(m('inputSchema.validation.objectKeysInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidKeys: invalidKeys.join(','),
+                        pattern: patternKey,
+                    }));
+                }
+            }
+            if (patternValue) {
+                const check = new RegExp(patternValue);
+                const invalidKeys: any[] = [];
+                Object.keys(value).forEach((key) => {
+                    const propertyValue = value[key];
+                    if (typeof propertyValue !== 'string' || !check.test(propertyValue)) invalidKeys.push(key);
+                });
+                if (invalidKeys.length) {
+                    fieldErrors.push(m('inputSchema.validation.objectValuesInvalid', {
+                        rootName: 'input',
+                        fieldKey: property,
+                        invalidKeys: invalidKeys.join(','),
+                        pattern: patternValue,
+                    }));
+                }
+            }
+        }
+        if (fieldErrors.length > 0) {
+            const message = fieldErrors.join(', ');
+            errors.push({ fieldKey: property, message });
+        }
+    });
+
+    return errors;
+}
